==> Scripts/1. IC Data Wrangling & Subsetting.ipynb <==
## Importing Libraries

# Import libraries
import pandas as pd
import numpy as np
import os

## Importing Data

path = r'/Users/vanessanunezpenas/Documents/Data Analytics Program - CareerFoundry/II Data Immersion 1 - Preparing and Analyzing Data/A4 - Python Fundamentals/2024.01 Instacart Basket Analysis'

path

# Importing orders.csv

df_ords = pd.read_csv(os.path.join(path, '02 Data', 'Original Data', 'orders.csv'), index_col = False)

# Importing products.csv

df_prods = pd.read_csv(os.path.join(path, '02 Data', 'Original Data', 'products.csv'), index_col = False)

## Data Wrangling Procedures

# Dropping 'eval_set' column from orders.csv because it's not relevant

df_ords.drop(columns = ['eval_set'])

df_ords['days_since_prior_order'].value_counts(dropna = False)

# Renaming 'order_dow' Column

df_ords.rename(columns = {'order_dow' : 'orders_day_of_week'}, inplace = True)

df_ords.head()

# Changing 'order_id' Data Type to 'str'

df_ords['order_id'] = df_ords['order_id'].astype('str')

df_ords['order_id'].dtype

# Importing departments.csv

df_dep = pd.read_csv(os.path.join(path, '02 Data', 'Original Data', 'departments.csv'), index_col = False)

df_dep.head()

# Transposing df_dep

df_dep.T

df_dep_t = df_dep.T
df_dep_t.reset_index()

# New header for df_dep_t

new_header = df_dep_t.iloc[0]

df_dep_t_new = df_dep_t[1:]

df_dep_t_new

df_dep_t_new.columns = new_header
df_dep_t_new

## Data Dictionaries

# Creating a data dictionary from 'department_id'

data_dict = df_dep_t_new.to_dict('index')
data_dict

df_prods.head()

print(data_dict.get('19'))

## Subsetting

# Creating a subset for df_prods containing only data from the snacks department

df_snacks =  df_prods[df_prods['department_id']==19]
df_snacks.head()

# Second way using loc function

df_snacks_2 = df_prods.loc[df_prods['department_id'] == 19]
df_snacks_2.head()

# Third way

df_snacks_3 = df_prods.loc[df_prods['department_id'].isin([19])]
df_snacks_3.head()

## Answering some Key Questions

# Changing 'user_id' Data Type to 'str'

df_ords['user_id'] = df_ords['user_id'].astype('str')

df_ords['user_id'].dtype

# Renaming 'order_hour_of_day'

df_ords.rename(columns = {'order_hour_of_day' : 'order_time'})

# what the busiest hour is for placing orders

df_ords['order_hour_of_day'].value_counts()

--> The busiest hour for placing orders is at 10:00.

# meaning behind a value of 4 in the "department_id" column within the df_prods

print(data_dict.get('4'))

# breakfast item sales

df_breakfast = df_prods[df_prods['department_id']==14] # To Create the subset
df_breakfast

# dinner parties item sales

# Subset for alcohol, deli, beverages, and meat/seafood
df_dinner_parties = df_prods.loc[df_prods['department_id'].isin([5, 20, 7, 12])]
df_dinner_parties

# Count of rows in df_dinner_parties
df_dinner_parties.shape

# Information about the customer with 'user_id' of '1'

df_ords['user_id']=='1'

df_ords_user_id_1 = df_ords[df_ords['user_id']=='1']
df_ords_user_id_1

# details about this user_id '1' behavior

df_ords_user_id_1.describe()

## Exporting the DataFrame

df_ords.to_csv(os.path.join(path, '02 Data','Prepared Data', 'orders_wrangled.csv'))

df_dep_t_new.to_csv(os.path.join(path, '02 Data', 'Prepared Data', 'departments_wrangled.csv'))
