==> Scripts/2. IC Data Consistency Check.ipynb <==
## Importing Libraries

# Import libraries
import pandas as pd
import numpy as np
import os

## Importing Data

path = path = r'/Users/vanessanunezpenas/Documents/Data Analytics Program - CareerFoundry/II Data Immersion 1 - Preparing and Analyzing Data/A4 - Python Fundamentals/2024.01 Instacart Basket Analysis'
path

# Importing product.csv

df_prods = pd.read_csv(os.path.join(path, '02 Data', 'Original Data', 'products.csv'), index_col=False)

# Importing orders_wrangled.csv

df_ords = pd.read_csv(os.path.join(path, '02 Data', 'Prepared Data', 'orders_wrangled.csv'), index_col=False)

## Data Consistency Check - products.csv


df_ords.describe()

### Mixed-Type Data

# Create a dataframe to practice

df_test = pd.DataFrame()

# Create a mixed type column

df_test['mix'] = ['a', 'b', 1, True]

df_test

# Check for mixed types

for col in df_test.columns.tolist():
  weird = (df_test[[col]].applymap(type) != df_test[[col]].iloc[0].apply(type)).any(axis = 1)
  if len (df_test[weird]) > 0:
    print (col)

# Fixing the inconsistency

df_test['mix'] = df_test['mix'].astype('str')

# 2nd Check for mixed types

for col in df_test.columns.tolist():
  weird = (df_test[[col]].applymap(type) != df_test[[col]].iloc[0].apply(type)).any(axis = 1)
  if len (df_test[weird]) > 0:
    print (col)

### Missing Values

# Finding missing values in products.csv

df_prods.isnull().sum()

# Subset containing only the missing values

df_nan = df_prods[df_prods['product_name'].isnull() == True]
df_nan

# Addressing missing values

df_prods.shape # to check the proportion of the missing values

# Create a new df without the missing values

df_prods_clean = df_prods[df_prods['product_name'].isnull() == False]

df_prods_clean.shape

### Duplicates

# Finding duplicates

df_dups = df_prods_clean[df_prods_clean.duplicated()]
df_dups

# Create a new df without the duplicates

df_prods_no_dups = df_prods_clean.drop_duplicates()

df_prods_no_dups.shape

## Data Consistency Checks - orders_wrangled.csv

df_ords.describe()

The values seem normal, but there is a column named Unnamed:0 with the index of the dataframe that we should delete. 

df_ords.head()

# Deleting the 'Unnamed: 0' column

df_ords = df_ords.drop(columns = ['Unnamed: 0'])

# Deleting again 'eval_set' column

df_ords = df_ords.drop(columns = ['eval_set'])

df_ords.head()

### Mixed-Type Data

# Check for mixed types

for col in df_test.columns.tolist():
  weird = (df_test[[col]].applymap(type) != df_test[[col]].iloc[0].apply(type)).any(axis = 1)
  if len (df_test[weird]) > 0:
    print (col)

There is no inconsistencies

### Missing Values

# Checking missing values

df_ords.isnull().sum()

There are 206209 missing values in column 'days_since_prior_order', probably because they refer to the first order of a new user. That's why we shouldn't delete or impute this missing values. One solution would be creating a new value as flag, but this should be a string and would mean creating mixed-type data in the column. Considering these factors, maybe the best option here is to leave the missing values untouched.

# Creating a subset to view the missing values

df_ords_nan = df_ords[df_ords['days_since_prior_order'].isnull() == True]

df_ords_nan.head()

df_ords_nan.shape

### Duplicates

# Checking for duplicates

df_dups = df_ords[df_ords.duplicated()]

df_dups

No duplicates

### Exporting the final, cleaned .csv

# df_prods

df_prods_clean.to_csv(os.path.join(path,'02 Data','Prepared Data', 'products_checked.csv'))

# df_ords

df_ords.to_csv(os.path.join(path, '02 Data', 'Prepared Data', 'orders_checked.csv'))
