==> Scripts/3. IC Combining & Exporting Data.ipynb <==
## Importing Libraries and Data Sets

import pandas as pd
import numpy as np
import os

path = path = r'/Users/vanessanunezpenas/Documents/Data Analytics Program - CareerFoundry/II Data Immersion 1 - Preparing and Analyzing Data/A4 - Python Fundamentals/2024.01 Instacart Basket Analysis'
path

# Import the orders_products_combined.pkl

orders_products_combined = pd.read_pickle(os.path.join(path, '02 Data', 'Prepared Data', 'orders_products_combined.pkl'))

# Check the shape

orders_products_combined.shape

# Import products_checked.csv

products = pd.read_csv(os.path.join(path, '02 Data', 'Prepared Data', 'products_checked.csv'))

products.shape

## Combining the two Data Sets

products.head()

orders_products_combined.head()

# Deleting the unnecessary Unnamed column

products = products.drop(columns = ['Unnamed: 0'])

products.head()

Q5: Both Data Sets have in common the column 'product_id', although we can use this column as key (common identifier). However, they don't have the same shape, so the best option to combine both is merging the data 

## To avoid a ValueError (Cannot use name of an existing column for indicator column), we need to delete the _merge column

orders_products_combined = orders_products_combined.drop(columns = '_merge')

# Merging the data

ords_prods_merge = orders_products_combined.merge(products, on = 'product_id', indicator = True)

# Confirming the results of the merge

ords_prods_merge.head()

ords_prods_merge['_merge'].value_counts()

ords_prods_merge.shape

## Exporting the dataframe

# Considering the size of the newly created dataframe, the best format to export the data is pickle

ords_prods_merge.to_pickle(os.path.join(path, '02 Data','Prepared Data', 'ords_prods_merge.pkl'))
