==> Scripts/4. IC Deriving New Variables.ipynb <==
import pandas as pd
import numpy as np
import os

path = r'/Users/vanessanunezpenas/Documents/Data Analytics Program - CareerFoundry/II Data Immersion 1 - Preparing and Analyzing Data/A4 - Python Fundamentals/2024.01 Instacart Basket Analysis'
path

ords_prods_merge = pd.read_pickle(os.path.join(path, '02 Data', 'Prepared Data', 'ords_prods_merge.pkl'))

# Create a subset of the dataframe with the first one million rows

df = ords_prods_merge[:1000000]

df.shape

## If-Statements with User-Defined Functions

# Define a function to flag products according to price

def price_label(row):

  if row['prices'] <= 5:
    return 'Low-range product'
  elif (row['prices'] > 5) and (row['prices'] <= 15):
    return 'Mid-range product'
  elif row['prices'] > 15:
    return 'High range'
  else: return 'Not enough data'

# Apply the function

df['price_range'] = df.apply(price_label, axis=1)

df['price_range'].value_counts(dropna = False)

df['prices'].max()

## If-Statements with the loc() Function

ords_prods_merge.loc[ords_prods_merge['prices'] > 15, 'price_range_loc'] = 'High-range product'

ords_prods_merge.loc[(ords_prods_merge['prices'] <= 15) & (ords_prods_merge['prices'] > 5), 'price_range_loc'] = 'Mid-range product'

ords_prods_merge.loc[ords_prods_merge['prices'] <= 5, 'price_range_loc'] = 'Low-range product'

ords_prods_merge['price_range_loc'].value_counts(dropna = False)

## If-Statements with For-Loops

ords_prods_merge['orders_day_of_week'].value_counts(dropna = False)

result = []

for value in ords_prods_merge["orders_day_of_week"]:
  if value == 0:
    result.append("Busiest day")
  elif value == 4:
    result.append("Least busy")
  else:
    result.append("Regularly busy")

ords_prods_merge['busiest_day'] = result

ords_prods_merge['busiest_day'].value_counts(dropna = False)

## Answering some key Questions

# checking the new columns “price_label” and “busiest_day”

ords_prods_merge.columns

# create a new the 'busiest_days' column

result_2 = []

for value in ords_prods_merge["orders_day_of_week"]:
  if value == 0 or value == 1:
    result_2.append("Busiest days")
  elif value == 4 or value == 3:
    result_2.append("Slowest days")
  else:
    result_2.append("Regularly busy")

ords_prods_merge['busiest_days'] = result_2

# Check the values of this new column for accuracy

ords_prods_merge['busiest_days'].value_counts(dropna = False)

ords_prods_merge.head()

After flagging the the two busiest days (Saturday and Sunday) and the two least busiest days (Wednesday and Tuesday), the most frequent value is "Regularly busy" (12,916,527), followed by "Busiest days" (11,864,860) and "Slowest days" (7,624,654). 

# identify the busiest hours of the day

ords_prods_merge['order_hour_of_day'].value_counts(dropna = False)

We will use an If-Statement with For-Loop to label busiest_period_of_day:
    - 'Most orders': >= 2 millionen orders/hour;
    - 'Average orders': 1-2 millionen order/hour;
    - 'Fewest orders': <= 1 million orders/hour

hour_counts = ords_prods_merge['order_hour_of_day'].value_counts(dropna = False)

result_3 = []

for hour in ords_prods_merge['order_hour_of_day']:
    count = hour_counts[hour]
    if count >= 2000000:
        result_3.append("Most orders")
    elif count <= 1000000:
        result_3.append("Fewest orders")
    else:
        result_3.append("Average orders")

ords_prods_merge['busiest_period_of_day'] = result_3

# Frequency of the new 'busiest_period_of_day'

ords_prods_merge['busiest_period_of_day'].value_counts(dropna = False)

# Exporting the dataframe as a pickle file

ords_prods_merge.to_pickle(os.path.join(path, '02 Data','Prepared Data', 'ords_prods_merge_derived.pkl'))
