==> Scripts/5. IC Groupping Data & Aggregating Variables.ipynb <==
import pandas as pd
import numpy as np
import os

path = r'/Users/vanessanunezpenas/Documents/Data Analytics Program - CareerFoundry/II Data Immersion 1 - Preparing and Analyzing Data/A4 - Python Fundamentals/2024.01 Instacart Basket Analysis'
path

ords_prods_merge = pd.read_pickle(os.path.join(path, '02 Data', 'Prepared Data', 'ords_prods_merge_derived.pkl'))

ords_prods_merge.shape

ords_prods_merge.head(10)

## Find the aggregated mean of the “order_number” column grouped by “department_id”

ords_prods_merge.groupby('department_id').agg({'order_number': ['mean']})

## Analyze the result

Now that we have the entire dataframe, we can see all the departments and not only a few. The order_number mean has changed for those departments that we could see in the subset because it is calculated with more than 32,000,000 records and not only 1,000,000. Considering the entire dataset, the department with the highest mean is the ID 21 (missing), and the lowest the ID 5 (alcohol). This means that a high number of order have no department assigned. The second higuest mean is the ID 10, corresponisng to bulk department. 

## Create a loyalty flag for existing customers

ords_prods_merge['max_order'] = ords_prods_merge.groupby(['user_id'])['order_number'].transform(np.max)

ords_prods_merge.head(15)

ords_prods_merge.loc[ords_prods_merge['max_order'] > 40, 'loyalty_flag'] = 'Loyal customer'

ords_prods_merge.loc[(ords_prods_merge['max_order'] <= 40) & (ords_prods_merge['max_order'] > 10), 'loyalty_flag'] = 'Regular customer'

ords_prods_merge.loc[ords_prods_merge['max_order'] <= 10, 'loyalty_flag'] = 'New customer'

ords_prods_merge['loyalty_flag'].value_counts(dropna = False)

 ords_prods_merge[['user_id', 'loyalty_flag', 'order_number']].head(15)

## Determine whether the prices of products purchased by loyal customers differ from those purchased by regular or new customers

ords_prods_merge.groupby('loyalty_flag').agg({'prices':['mean','min','max']})

New customers spend more than regular and loyal customers. Actually, the loyal customers spend in average less than any other customer. 

## Target different types of spenders

ords_prods_merge['average_spending'] = ords_prods_merge.groupby(['user_id'])['prices'].transform(np.mean)

ords_prods_merge.head(20)

ords_prods_merge.loc[ords_prods_merge['average_spending'] < 10, 'spending_flag'] = 'Low spender'

ords_prods_merge.loc[ords_prods_merge['average_spending'] >= 10, 'spending_flag'] = 'High spender'

ords_prods_merge['spending_flag'].value_counts(dropna = False)

## Determine frequent versus non-frequent customers

ords_prods_merge['regularity'] = ords_prods_merge.groupby(['user_id'])['days_since_prior_order'].transform(np.median)

ords_prods_merge.loc[ords_prods_merge['regularity'] > 20, 'frequency_flag'] = 'Non-frequent customer'

ords_prods_merge.loc[(ords_prods_merge['regularity'] > 10) & (ords_prods_merge['regularity'] <= 20), 'frequency_flag'] = 'Regular customer'

ords_prods_merge.loc[ords_prods_merge['regularity'] <= 10, 'frequency_flag'] = 'Frequent customer'

ords_prods_merge['frequency_flag'].value_counts(dropna = False)

## *Bonus: cleaning data

# Exploratory checks

ords_prods_merge.loc[ords_prods_merge['prices'] > 100]

# Mark the outliers as missing

ords_prods_merge.loc[ords_prods_merge['prices'] >100, 'prices'] = np.nan

ords_prods_merge['prices'].max()

## Export the dataframe as pickle file

ords_prods_merge.to_pickle(os.path.join(path, '02 Data','Prepared Data', 'ords_prods_merge_final.pkl'))
