==> Scripts/6.  IC Preparing Data Visualizations.ipynb <==
## Import Libraries and customer data set

import pandas as pd
import numpy as np
import os
import matplotlib.pyplot as plt
import seaborn as sns
import scipy

path = r'/Users/vanessanunezpenas/Documents/Data Analytics Program - CareerFoundry/II Data Immersion 1 - Preparing and Analyzing Data/A4 - Python Fundamentals/2024.01 Instacart Basket Analysis'
path

customers = pd.read_csv(os.path.join(path, '02 Data', 'Original Data', 'customers.csv'))

customers.head(15)

customers.shape

customers.dtypes

## Wrangle the data

# Renaming columns

columns_rename = {'Surnam': 'Surname',
                'STATE': 'State',
                'fam_status': 'familial_status'}
customers.rename(columns = columns_rename, inplace = True)

# Changing 'user_id' data type to str

customers['user_id'] = customers['user_id'].astype('str')

# M y first intention was to drop the column 'n_dependants', but I have checked that we need it in Part 2

## Data quality and consistency checks

# Check for mixed types

for col in customers.columns.tolist():
  weird = (customers[[col]].applymap(type) != customers[[col]].iloc[0].apply(type)).any(axis = 1)
  if len (customers[weird]) > 0:
    print (col)

# Fixing the inconsistency

customers['First Name'] = customers['First Name'].astype('str')

# 2nd Check for mixed types

for col in customers.columns.tolist():
  weird = (customers[[col]].applymap(type) != customers[[col]].iloc[0].apply(type)).any(axis = 1)
  if len (customers[weird]) > 0:
    print (col)

# Finding missing values in products.csv

customers.isnull().sum()

# Finding duplicates

customers_dups = customers[customers.duplicated()]
customers_dups

customers.head()

customers.describe()

## Combine the data

# Importing ords_prods_merge.pkl

ords_prods_merge = pd.read_pickle(os.path.join(path, '02 Data', 'Prepared Data', 'ords_prods_merge_final.pkl'))

ords_prods_merge.shape

ords_prods_merge.columns

ords_prods_merge.head()

# Checking 'user_id' data type in ords_prods_merge

ords_prods_merge['user_id'].dtype

# Changing 'user_id' data type to str

ords_prods_merge['user_id'] = ords_prods_merge['user_id'].astype('str')

# To avoid a ValueError (Cannot use name of an existing column for indicator column), we need to delete the _merge column
# I will drop as well other columns we don't need anymore, like 'max_order'

ords_prods_merge = ords_prods_merge.drop(columns = ['_merge', 'max_order', 'average_spending', 'regularity'])

# Merging the data

orders_products_all = ords_prods_merge.merge(customers, on = 'user_id', indicator = True)

# Confirming the results of the merge

orders_products_all.head()

orders_products_all['_merge'].value_counts()

orders_products_all.shape

## Export the new dataframe as a pickle file

orders_products_all.to_pickle(os.path.join(path, '02 Data','Prepared Data', 'orders_products_all.pkl'))
